Move transition.tsx to lib/component
